# server/cmd/api/docs/swagger.yaml
definitions:
  request.Login:
    properties:
      email:
        type: string
      password:
        maxLength: 30
        minLength: 5
        type: string
      phone:
        maxLength: 10
        minLength: 10
        type: string
      user_name:
        maxLength: 15
        minLength: 3
        type: string
    required:
    - password
    type: object
  request.RefreshToken:
    properties:
      refresh_token:
        minLength: 10
        type: string
    type: object
  request.UserSignUp:
    properties:
      age:
        type: integer
      confirm_password:
        type: string
      email:
        type: string
      first_name:
        maxLength: 50
        minLength: 2
        type: string
      last_name:
        maxLength: 50
        minLength: 1
        type: string
      password:
        type: string
      user_name:
        maxLength: 15
        minLength: 3
        type: string
    required:
    - age
    - confirm_password
    - email
    - first_name
    - last_name
    - password
    - user_name
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /api/auth/login:
    post:
      description: API for user to login with email | phone | user_name with password
      operationId: UserLogin
      parameters:
      - description: Login Details
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      responses:
        "200":
          description: Successfully logged in
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: User not exist with given login credentials
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: User blocked by admin
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to login
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Login with password (User)
      tags:
      - User Authentication
  /api/auth/renew-access-token:
    post:
      description: API for user to renew access token using refresh token
      operationId: UserRenewAccessToken
      parameters:
      - description: Refresh token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RefreshToken'
      responses:
        "200":
          description: Successfully generated access token using refresh token
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Refresh token blocked
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: No session found for the given refresh token
          schema:
            $ref: '#/definitions/response.Response'
        "410":
          description: Refresh token expired
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed generate access token
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Renew Access Token (User)
      tags:
      - User Authentication
  /api/auth/signup:
    post:
      description: API for user to register a new account
      operationId: UserSignUp
      parameters:
      - description: Input Fields
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.UserSignUp'
      responses:
        "200":
          description: Successfully Account Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: A verified user already exist with given user credentials
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to signup
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Signup (User)
      tags:
      - User Authentication
swagger: "2.0"